requirements.txt
aenum==3.1.15
aiobotocore==2.12.3
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.4
arrow-pd-parser==1.3.9
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
awswrangler==3.7.2
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==42.0.8
dataengineeringutils3==1.4.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
fsspec==2024.3.1
gremlinpython==3.7.1
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mojap-metadata @ git+https://github.com/moj-analytical-services/mojap-metadata@8ebaa52f3841fc3669c49039606200e35eda23b0
moto==5.0.9
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.23.4
openpyxl==3.0.10
opensearch-py==2.5.0
packaging==24.0
pandas==1.5.1
parameterized==0.7.5
parso==0.8.4
pexpect==4.9.0
pg8000==1.31.1
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
progressbar2==4.4.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydbtools==5.5.16
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
pytest==7.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redshift-connector==2.0.918
referencing==0.34.0
requests==2.31.0
requests-aws4auth==1.2.3
responses==0.25.3
rpds-py==0.18.0
s3fs==2024.3.1
s3transfer==0.10.1
scramp==1.4.4
six==1.16.0
smart-open==5.2.1
soupsieve==2.5
sql_metadata==2.10.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.18.1
